refactor: restructure Surah components for improved readability and maintainability
